Changed good vertex filter cut |z| < 18 cm to |z| < 24 cm, see https://hypernews.cern.ch/HyperNews/CMS/get/physics-validation/986/1.html

+++ ElectroWeakAnalysis/MultiBosons/python/Skimming/goodCollisionDataSequence_cff.py
@@ -25,7 +25,7 @@ noScraping = cms.EDFilter("FilterOutScraping",
 primaryVertexFilter = cms.EDFilter("GoodVertexFilter",
   vertexCollection = cms.InputTag('offlinePrimaryVertices'),
   minimumNDOF = cms.uint32(4),
-  maxAbsZ = cms.double(18),
+  maxAbsZ = cms.double(24),
   maxd0 = cms.double(2)
 )
